+++ TrackPropagation/RungeKutta/src/PathToPlane2Order.h
@@ -29,7 +29,7 @@ public:
 
     /// the position and momentum are local in the FieldFrame;
     /// the plane is in the global frame
-    std::pair<bool,double> operator()( const HessianPlane<float>& plane, 
+    std::pair<bool,double> operator()( const HessianPlane<double>& plane, 
 				       const Vector3D& position,
 				       const Vector3D& momentum,
 				       float charge,
